lpt_drv: added source to old lpt driver
Import source to older linux 2.4 device driver for parallel port
cameras.  This code was provided to customers on the SBIG web site
as lpt_drv.tar.gz but has since been removed.  It was provided to
me by Tim Puckett, VP of sales at Cyanogen/SBIG.  Hopefully we can
get it ported to a modern Linux kernel.

Stracing the current sbigudrv sdk, it would appear that the code
for communicating with this driver remains in place although it surely
has not been tested on Linux for some time.  It is unclear whether an
older version of the sdk from the same era as the driver (2005-2008?)
will need to be employed.

The original directory structure was flattend from:
  Makefile
  module/sbig/<source code>
  release/module/sbig/<module binary>
and build products were excluded.