lpt_drv: use HZ instead of gLptHz
Problem: driver hardwires several HZ values and attempts
to time elapsed jiffies to determine which it is, then uses
that as the basis for timeouts.  My HZ is 250 but that was
not one of the expected values, so it used 1.

The above code is fragile, and I'm not sure I understand
what it accomplished.  Switch back to using the defined HZ
value rather than the calculated one.