driver: [cleanup] drop dead code
Problem: The driver contains a considerable amount of code
from SBIG's internal multi-platform build tree, as well as a
complete copy of sbigudrv.h which defines the API for the universal
driver SDK.  This makes the code hard to follow.

Since the driver no longer gets updates from SBIG/DL, it should be
safe to cull the code that does not pertain directly to this driver.

As a result some header files were removed, and some with very
little left in them after unused bits were removed were combined
with other source files.