driver: printk with correct struct device
Problem: printk macros are using the 'struct device'
belonging to the parallel port to call dev_{level}().
This results in console messages prefixed with e.g..
"parport_pc 0000:03:00.0" and it's hard to tell that
it's from this driver.

Capture the 'struct device' returned by device_create()
and use that. Now prefix is e.g. "sbiglpt sbiglpt0".