driver: clarify license on DL-owned files
Problem: the original SBIG release of the parallel port
driver source code did not specify a license.

Diffraction Limited, the current owner of this code,
has clarified that they wish it to be released under
the GPL-2.0 license, for compatibility with the Linux
kernel.

Update the copyright notice and SPDX license tags
accordingly.

These changes were approved by:
  Douglas B. George <dgeorge@diffractionlimted.com>