build: tell kbuild about external Module.ksymvers
Problem: build fails with undefined parport symobols
if parport was built outside of the kernel tree,
e.g. on raspbian where it is built within pi-parport.

Ensure that KBUILD_EXTRA_SYMBOLS is carried through to
the inner make invocation if set on the command line.

When building with pi-parport, use something like:
  make KBUILD_EXTRA_SYMBOLS=$(pwd)/../../pi-parport/driver/Module.symvers